Elk-ioc/src/test/java/com/thoughtworks/elk/injection/test/ConstructorInjectionTest.java
package com.thoughtworks.elk.injection.test;

import com.thoughtworks.elk.container.ElkContainer;
import com.thoughtworks.elk.injection.Injection;
import com.thoughtworks.elk.injection.ConstructorInjection;
import com.thoughtworks.elk.movie.Director;
import com.thoughtworks.elk.movie.Hollywood;
import com.thoughtworks.elk.movie.Titanic;
import org.junit.Test;

import java.lang.reflect.InvocationTargetException;

import static org.hamcrest.core.IsNull.notNullValue;
import static org.junit.Assert.assertThat;

public class ConstructorInjectionTest {
    @Test
    public void should_build_a_bean_with_dependencies()
            throws IllegalAccessException, InstantiationException, InvocationTargetException {
        Injection injection = new ConstructorInjection();
        ElkContainer elkContainer = new ElkContainer(injection);
        elkContainer.register(Titanic.class);
        elkContainer.register(Hollywood.class);
        elkContainer.register(Director.class);

        Director director = injection.buildBeanWithDependencies(Director.class, elkContainer);

        assertThat(director, notNullValue());
    }
}
